+++ mex/common/models/contact_point.py
@@ -11,15 +11,15 @@ from mex.common.models.base.preview_item import PreviewItem
 from mex.common.models.base.rules import (
     AdditiveRule,
     PreventiveRule,
-    PublishingRule,
     RuleSet,
     SubtractiveRule,
+    WorkflowRule,
 )
 from mex.common.types import (
     ExtractedContactPointIdentifier,
     MergedContactPointIdentifier,
     MergedPrimarySourceIdentifier,
-    PublishingStatus,
+    PublishingTarget,
 )
 
 EmailStr = Annotated[
@@ -227,16 +227,16 @@ class PreventiveContactPoint(_Stem, PreventiveRule):
     email: list[MergedPrimarySourceIdentifier] = []
 
 
-class PublishingContactPoint(_Stem, PublishingRule):
+class WorkflowContactPoint(_Stem, WorkflowRule):
     """Rule to prevent publishing of merged contact point items."""
 
     entityType: Annotated[
-        Literal["PublishingContactPoint"], Field(alias="$type", frozen=True)
-    ] = "PublishingContactPoint"
-    status: Annotated[
-        PublishingStatus | None,
-        Field(description="Indicates if the merged item should NOT be published."),
-    ] = None
+        Literal["WorkflowContactPoint"], Field(alias="$type", frozen=True)
+    ] = "WorkflowContactPoint"
+    forbiddenPublishingTarget: Annotated[
+        list[PublishingTarget],
+        Field(description="Targets to which the merged item should NOT be published."),
+    ] = []
 
 
 class _BaseRuleSet(_Stem, RuleSet):
@@ -245,7 +245,7 @@ class _BaseRuleSet(_Stem, RuleSet):
     additive: AdditiveContactPoint = AdditiveContactPoint()
     subtractive: SubtractiveContactPoint = SubtractiveContactPoint()
     preventive: PreventiveContactPoint = PreventiveContactPoint()
-    publishing: PublishingContactPoint = PublishingContactPoint()
+    workflow: WorkflowContactPoint = WorkflowContactPoint()
 
 
 class ContactPointRuleSetRequest(_BaseRuleSet):