+++ mex/common/models/resource.py
@@ -11,9 +11,9 @@ from mex.common.models.base.preview_item import PreviewItem
 from mex.common.models.base.rules import (
     AdditiveRule,
     PreventiveRule,
-    PublishingRule,
     RuleSet,
     SubtractiveRule,
+    WorkflowRule,
 )
 from mex.common.types import (
     AccessRestriction,
@@ -36,7 +36,7 @@ from mex.common.types import (
     MergedPrimarySourceIdentifier,
     MergedResourceIdentifier,
     PersonalData,
-    PublishingStatus,
+    PublishingTarget,
     ResourceCreationMethod,
     ResourceTypeGeneral,
     Text,
@@ -927,16 +927,16 @@ class PreventiveResource(_Stem, PreventiveRule):
     wasGeneratedBy: list[MergedPrimarySourceIdentifier] = []
 
 
-class PublishingResource(_Stem, PublishingRule):
+class WorkflowResource(_Stem, WorkflowRule):
     """Rule to prevent publishing of merged resource items."""
 
     entityType: Annotated[
-        Literal["PublishingResource"], Field(alias="$type", frozen=True)
-    ] = "PublishingResource"
-    status: Annotated[
-        PublishingStatus | None,
-        Field(description="Indicates if the merged item should NOT be published."),
-    ] = None
+        Literal["WorkflowResource"], Field(alias="$type", frozen=True)
+    ] = "WorkflowResource"
+    forbiddenPublishingTarget: Annotated[
+        list[PublishingTarget],
+        Field(description="Targets to which the merged item should NOT be published."),
+    ] = []
 
 
 class _BaseRuleSet(_Stem, RuleSet):
@@ -945,7 +945,7 @@ class _BaseRuleSet(_Stem, RuleSet):
     additive: AdditiveResource = AdditiveResource()
     subtractive: SubtractiveResource = SubtractiveResource()
     preventive: PreventiveResource = PreventiveResource()
-    publishing: PublishingResource = PublishingResource()
+    workflow: WorkflowResource = WorkflowResource()
 
 
 class ResourceRuleSetRequest(_BaseRuleSet):