+++ mex/common/types/__init__.py
@@ -35,7 +35,7 @@ from mex.common.types.identifier import (
 from mex.common.types.identity import IdentityProvider
 from mex.common.types.link import Link, LinkLanguage
 from mex.common.types.path import AssetsPath, PathWrapper, WorkPath
-from mex.common.types.publishing_status import PublishingStatus
+from mex.common.types.publishing_target import PublishingTarget
 from mex.common.types.sink import Sink
 from mex.common.types.temporal_entity import (
     CET,
@@ -143,7 +143,7 @@ __all__ = (
     "MergedVariableIdentifier",
     "PathWrapper",
     "PersonalData",
-    "PublishingStatus",
+    "PublishingTarget",
     "ResourceCreationMethod",
     "ResourceTypeGeneral",
     "Sink",